+++ backend/src/modules/communication/communication.module.ts
@@ -1,25 +1,25 @@
 import { Module } from '@nestjs/common';
 
-import BoardsModule from 'src/modules/boards/boards.module';
+import BoardsModule from 'modules/boards/boards.module';
 import {
-  CommunicationGateAdapter,
-  ChatHandler,
-  ConversationsHandler,
-  UsersHandler,
-  ExecuteCommunication,
-} from 'src/modules/communication/communication.providers';
-import { SlackExecuteCommunicationService } from 'src/modules/communication/services/slack-execute-communication.service';
+	CommunicationGateAdapter,
+	ChatHandler,
+	ConversationsHandler,
+	UsersHandler,
+	ExecuteCommunication
+} from 'modules/communication/communication.providers';
+import { SlackExecuteCommunicationService } from 'modules/communication/services/slack-execute-communication.service';
 
 @Module({
-  imports: [BoardsModule],
-  providers: [
-    SlackExecuteCommunicationService,
-    CommunicationGateAdapter,
-    ChatHandler,
-    ConversationsHandler,
-    UsersHandler,
-    ExecuteCommunication,
-  ],
-  exports: [SlackExecuteCommunicationService],
+	imports: [BoardsModule],
+	providers: [
+		SlackExecuteCommunicationService,
+		CommunicationGateAdapter,
+		ChatHandler,
+		ConversationsHandler,
+		UsersHandler,
+		ExecuteCommunication
+	],
+	exports: [SlackExecuteCommunicationService]
 })
 export class CommunicationModule {}
