+++ backend/src/modules/communication/dto/team.dto.ts
@@ -1,20 +1,20 @@
-import { BoardRoles } from 'src/modules/communication/dto/types';
-import { UserDto } from 'src/modules/communication/dto/user.dto';
+import { BoardRoles } from 'modules/communication/dto/types';
+import { UserDto } from 'modules/communication/dto/user.dto';
 
 export class TeamDto {
-  name!: string;
+	name!: string;
 
-  normalName!: string;
+	normalName!: string;
 
-  boardId!: string;
+	boardId!: string;
 
-  channelId?: string;
+	channelId?: string;
 
-  type!: 'team' | 'sub-team';
+	type!: 'team' | 'sub-team';
 
-  for!: BoardRoles;
+	for!: BoardRoles;
 
-  participants!: UserDto[];
+	participants!: UserDto[];
 
-  participantsNotInvited?: string[];
+	participantsNotInvited?: string[];
 }
