+++ backend/src/modules/communication/handlers/chat-slack.handler.spec.ts
@@ -1,75 +1,67 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable @typescript-eslint/lines-between-class-members */
-import { ChatSlackHandler } from 'src/modules/communication/handlers/chat-slack.handler';
-import { CommunicationGateInterface } from 'src/modules/communication/interfaces/communication-gate.interface';
+import { ChatSlackHandler } from 'modules/communication/handlers/chat-slack.handler';
+import { CommunicationGateInterface } from 'modules/communication/interfaces/communication-gate.interface';
 
 const MakeSlackCommunicationGateAdapterStub = () => {
-  class SlackCommunicationGateAdapterStub
-    implements CommunicationGateInterface
-  {
-    addChannel(name: string): Promise<{ id: string; name: string }> {
-      throw new Error('Method not implemented.');
-    }
-    addUsersToChannel(
-      channelId: string,
-      usersIds: string[],
-    ): Promise<{ ok: boolean; fails?: string[] | undefined }> {
-      throw new Error('Method not implemented.');
-    }
-    getAllUsersByChannel(channelId: string): Promise<string[]> {
-      throw new Error('Method not implemented.');
-    }
-    getEmailByUserId(userId: string): Promise<string> {
-      throw new Error('Method not implemented.');
-    }
-    addMessageToChannel(channelId: string, message: string): Promise<boolean> {
-      return Promise.resolve(true);
-    }
-  }
-  return new SlackCommunicationGateAdapterStub();
+	class SlackCommunicationGateAdapterStub implements CommunicationGateInterface {
+		addChannel(name: string): Promise<{ id: string; name: string }> {
+			throw new Error('Method not implemented.');
+		}
+		addUsersToChannel(
+			channelId: string,
+			usersIds: string[]
+		): Promise<{ ok: boolean; fails?: string[] | undefined }> {
+			throw new Error('Method not implemented.');
+		}
+		getAllUsersByChannel(channelId: string): Promise<string[]> {
+			throw new Error('Method not implemented.');
+		}
+		getEmailByUserId(userId: string): Promise<string> {
+			throw new Error('Method not implemented.');
+		}
+		addMessageToChannel(channelId: string, message: string): Promise<boolean> {
+			return Promise.resolve(true);
+		}
+	}
+	return new SlackCommunicationGateAdapterStub();
 };
 
 describe('ChatSlackHandler', () => {
-  let handler: ChatSlackHandler;
-  const slackCommunicationGateAdapterStub =
-    MakeSlackCommunicationGateAdapterStub();
+	let handler: ChatSlackHandler;
+	const slackCommunicationGateAdapterStub = MakeSlackCommunicationGateAdapterStub();
 
-  beforeAll(async () => {
-    handler = new ChatSlackHandler(slackCommunicationGateAdapterStub);
-  });
+	beforeAll(async () => {
+		handler = new ChatSlackHandler(slackCommunicationGateAdapterStub);
+	});
 
-  it('should be defined', () => {
-    expect(handler).toBeDefined();
-  });
+	it('should be defined', () => {
+		expect(handler).toBeDefined();
+	});
 
-  it('should post a message by channel id', async () => {
-    const channelId = 'any_channel_id';
-    const message = 'any_message';
+	it('should post a message by channel id', async () => {
+		const channelId = 'any_channel_id';
+		const message = 'any_message';
 
-    const spy = jest.spyOn(
-      slackCommunicationGateAdapterStub,
-      'addMessageToChannel',
-    );
+		const spy = jest.spyOn(slackCommunicationGateAdapterStub, 'addMessageToChannel');
 
-    const result = await handler.postMessage(channelId, message);
+		const result = await handler.postMessage(channelId, message);
 
-    expect(result).toMatchObject({ ok: true, channel: 'any_channel_id' });
-    expect(spy).toHaveBeenCalledTimes(1);
-    spy.mockRestore();
-  });
+		expect(result).toMatchObject({ ok: true, channel: 'any_channel_id' });
+		expect(spy).toHaveBeenCalledTimes(1);
+		spy.mockRestore();
+	});
 
-  it('should throw error when adapter throws', async () => {
-    const channelId = 'any_channel_id';
-    const message = 'any_message';
+	it('should throw error when adapter throws', async () => {
+		const channelId = 'any_channel_id';
+		const message = 'any_message';
 
-    const spy = jest
-      .spyOn(slackCommunicationGateAdapterStub, 'addMessageToChannel')
-      .mockImplementation(() => Promise.reject(new Error('some error')));
+		const spy = jest
+			.spyOn(slackCommunicationGateAdapterStub, 'addMessageToChannel')
+			.mockImplementation(() => Promise.reject(new Error('some error')));
 
-    await expect(
-      handler.postMessage(channelId, message),
-    ).rejects.toThrowError();
-    expect(spy).toHaveBeenCalledTimes(1);
-    spy.mockRestore();
-  });
+		await expect(handler.postMessage(channelId, message)).rejects.toThrowError();
+		expect(spy).toHaveBeenCalledTimes(1);
+		spy.mockRestore();
+	});
 });
