+++ backend/src/modules/communication/handlers/conversations-slack.handler.ts
@@ -1,35 +1,31 @@
-import { CommunicationGateInterface } from 'src/modules/communication/interfaces/communication-gate.interface';
-import { ConversationsHandlerInterface } from 'src/modules/communication/interfaces/conversations.handler.interface';
-import { UserDto } from 'src/modules/communication/dto/user.dto';
+import { CommunicationGateInterface } from 'modules/communication/interfaces/communication-gate.interface';
+import { ConversationsHandlerInterface } from 'modules/communication/interfaces/conversations.handler.interface';
+import { UserDto } from 'modules/communication/dto/user.dto';
 
-export class ConversationsSlackHandler
-  implements ConversationsHandlerInterface
-{
-  constructor(
-    private readonly communicationGateAdapter: CommunicationGateInterface,
-  ) {}
+export class ConversationsSlackHandler implements ConversationsHandlerInterface {
+	constructor(private readonly communicationGateAdapter: CommunicationGateInterface) {}
 
-  async createChannel(name: string): Promise<{ name: string; id: string }> {
-    return this.communicationGateAdapter.addChannel(name);
-  }
+	async createChannel(name: string): Promise<{ name: string; id: string }> {
+		return this.communicationGateAdapter.addChannel(name);
+	}
 
-  async inviteUsersToChannel(
-    channelId: string,
-    users: UserDto[],
-  ): Promise<{ ok: boolean; channelId: string; fails?: string[] }> {
-    const usersIds = users
-      .filter((i) => typeof i.slackId === 'string')
-      .map((i) => i.slackId as string);
+	async inviteUsersToChannel(
+		channelId: string,
+		users: UserDto[]
+	): Promise<{ ok: boolean; channelId: string; fails?: string[] }> {
+		const usersIds = users
+			.filter((i) => typeof i.slackId === 'string')
+			.map((i) => i.slackId as string);
 
-    const { ok, fails } = await this.communicationGateAdapter.addUsersToChannel(
-      channelId,
-      usersIds,
-    );
+		const { ok, fails } = await this.communicationGateAdapter.addUsersToChannel(
+			channelId,
+			usersIds
+		);
 
-    return { ok, channelId, fails };
-  }
+		return { ok, channelId, fails };
+	}
 
-  async getUsersFromChannelSlowly(channelId: string): Promise<string[]> {
-    return this.communicationGateAdapter.getAllUsersByChannel(channelId);
-  }
+	async getUsersFromChannelSlowly(channelId: string): Promise<string[]> {
+		return this.communicationGateAdapter.getAllUsersByChannel(channelId);
+	}
 }
