+++ backend/src/modules/communication/interfaces/conversations.handler.interface.ts
@@ -1,12 +1,12 @@
-import { UserDto } from 'src/modules/communication/dto/user.dto';
+import { UserDto } from 'modules/communication/dto/user.dto';
 
 export interface ConversationsHandlerInterface {
-  createChannel(name: string): Promise<{ name: string; id: string }>;
+	createChannel(name: string): Promise<{ name: string; id: string }>;
 
-  inviteUsersToChannel(
-    channelId: string,
-    users: UserDto[],
-  ): Promise<{ ok: boolean; channelId: string; fails?: string[] }>;
+	inviteUsersToChannel(
+		channelId: string,
+		users: UserDto[]
+	): Promise<{ ok: boolean; channelId: string; fails?: string[] }>;
 
-  getUsersFromChannelSlowly(channelId: string): Promise<string[]>;
+	getUsersFromChannelSlowly(channelId: string): Promise<string[]>;
 }
