+++ backend/src/modules/communication/interfaces/execute-communication.interface.ts
@@ -1,6 +1,6 @@
-import { TeamDto } from 'src/modules/communication/dto/team.dto';
-import { BoardType } from 'src/modules/communication/dto/types';
+import { TeamDto } from 'modules/communication/dto/team.dto';
+import { BoardType } from 'modules/communication/dto/types';
 
 export interface ExecuteCommunicationInterface {
-  execute(board: BoardType): Promise<TeamDto[]>;
+	execute(board: BoardType): Promise<TeamDto[]>;
 }
